refactor gql types to separate file
